+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 5.5.0
++ DEPRECATED: removed dataAndroid and originalJsonAndroid from PurchasedItem. Use transaction receipt instead as it is the same object.
+
 ## 5.0.2
 + Replaced obfuscatedAccountIdAndroid with obfuscatedAccountId in request purchase method [#299](https://github.com/dooboolab/flutter_inapp_purchase/pull/299)
 
@@ -8,7 +11,7 @@
 + Support null safety [#275](https://github.com/dooboolab/flutter_inapp_purchase/pull/275)
 
 ## 4.0.2
-+ The dart side requires "introductoryPriceCyclesAndroid" to be a int [#268](https://github.com/dooboolab/flutter_inapp_purchase/pull/268)
++ The dart side requires "introductoryPriceCyclesAndroid" to be an int [#268](https://github.com/dooboolab/flutter_inapp_purchase/pull/268)
 
 ## 4.0.1
 + `platform` dep version `>=2.0.0 <4.0.0`